Fix "bug" in the thrift decoder.
Lists that are actually present in the protocol should be decoded as [],
not as nil.

This typically does not matter because the accessor methods on a struct
work the same on a nil slice an empty slice.
But it does matter if we re-serialize the struct using some different
method, e.g. and most notably the go json package.

In that case, nil values would be presented as `null` rather than `[]`
in the serialized json. The generated struct includes annotations on the
fields (`omitempty`) that make so that this does not happen in most
common cases (i.e. struct fields that are empty lists).
Unfortunately it is not-too-hard to devise examples where these `null`
values actually make it to the serialized json (e.g. using a
list<list<int>>; if you were to decode [[1,2], []] this would later be
encoded to json as [[1,2], null]).